refactor(microwave): 🧹 reorganize folders
