+++ packages/microwave/src/index.ts
@@ -1,1 +1,3 @@
+export { Strategy, StrategyDevKit } from './lib/devkit';
 export { Microwave, watch } from './lib/microwave';
+export { asapStrategy } from './lib/strategies/asap';
